week6/code-generation/requirements.txt
fastapi
uvicorn
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
redis